Uploading two new scripts
The following scripts are:
implementation.py: It allows the user to implement the functions described in matcher_saber.py. It also includes the location where the databases are (with their actual names). 

matcher_saber.py: This code includes the school class (the main element), the function to implement the class over a database and a function that allows the user to create the stata Do-file.